LayoutTests/fast/multicol/break-properties-expected.txt

PASS: 'break-before' is at (218, 8)
PASS: 'after-break' is at (428, 8)
PASS: 'no-break' is at (533, 8)
